requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.1.0
netcdf4>=1.6.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.20.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
PyYAML>=6.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
hydra-core>=1.3.0
omegaconf>=2.3.0
einops>=0.6.0
torch-geometric>=2.3.0
transformers>=4.30.0
timm>=0.9.0
